pyproject.toml
[project]
name = "image"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "mimetype>=0.1.5",
    "pillow>=12.0.0",
]
